022-burette/.magnum_opus/requirements.txt
appdirs
flask
flask-restful
flask-marshmallow
flask-sqlalchemy
flask-injector
marshmallow-sqlalchemy
sqlalchemy-utils
flask-cqlalchemy
cassandra-driver
